+++ R/plot_grid.R
@@ -19,19 +19,17 @@
 plot_grid <- function(regions = NULL, plot_object = "bigr_eff_plot.rds", 
                       results_dir = "results", target_date = NULL, ...) {
   
-plots <- 
+plots <- suppressWarnings(
   purrr::map(regions, function(region) {
     plot <- EpiNow::load_nowcast_result(plot_object, region, 
                                         date = target_date, results_dir) +
       ggplot2::labs(title = region %>% 
                       stringr::str_replace("-", " ") %>% 
                       stringr::str_to_title()) +
-      ggplot2::scale_x_date(date_breaks = "1 week", date_labels = "%b %d",
-                            limits = c(as.Date("2020-03-01"), as.Date(target_date))) +
-      ggplot2::coord_cartesian(ylim = c(0, 4))
+      ggplot2::scale_x_date(date_breaks = "1 week", date_labels = "%b %d")
     
     return(plot)
-  })
+  }))
   
   plot <- plots %>% 
     patchwork::wrap_plots() +